rename element to point
