+++ math/src/elliptic_curve/edwards/mod.rs
@@ -1,2 +1,2 @@
-pub mod element;
+pub mod point;
 pub mod traits;
