+++ math/src/elliptic_curve/mod.rs
@@ -1,6 +1,6 @@
 pub mod edwards;
 pub mod montgomery;
 /// Implementation of ProjectivePoint, a generic projective point in a curve.
-pub mod projective_point;
+pub mod point;
 pub mod short_weierstrass;
 pub mod traits;
