+++ math/src/elliptic_curve/montgomery/mod.rs
@@ -1,1 +1,1 @@
-pub mod element;
+pub mod point;
