+++ math/src/elliptic_curve/short_weierstrass/curves/bls12_377/curve.rs
@@ -1,5 +1,5 @@
 use super::field_extension::BLS12377PrimeField;
-use crate::elliptic_curve::short_weierstrass::element::ShortWeierstrassProjectivePoint;
+use crate::elliptic_curve::short_weierstrass::point::ShortWeierstrassProjectivePoint;
 use crate::elliptic_curve::traits::IsEllipticCurve;
 use crate::{
     elliptic_curve::short_weierstrass::traits::IsShortWeierstrass, field::element::FieldElement,
