+++ math/src/elliptic_curve/short_weierstrass/curves/bls12_381/curve.rs
@@ -1,4 +1,4 @@
-use crate::elliptic_curve::short_weierstrass::element::ShortWeierstrassProjectivePoint;
+use crate::elliptic_curve::short_weierstrass::point::ShortWeierstrassProjectivePoint;
 use crate::elliptic_curve::traits::IsEllipticCurve;
 use crate::{
     elliptic_curve::short_weierstrass::traits::IsShortWeierstrass, field::element::FieldElement,
