+++ math/src/elliptic_curve/short_weierstrass/curves/test_curve_2.rs
@@ -1,4 +1,4 @@
-use crate::elliptic_curve::short_weierstrass::element::ShortWeierstrassProjectivePoint;
+use crate::elliptic_curve::short_weierstrass::point::ShortWeierstrassProjectivePoint;
 use crate::elliptic_curve::traits::IsEllipticCurve;
 use crate::field::fields::u384_prime_field::{
     IsMontgomeryConfiguration, MontgomeryBackendPrimeField,
