+++ math/src/elliptic_curve/short_weierstrass/mod.rs
@@ -1,8 +1,8 @@
 /// Implementation of particular cases of elliptic curves.
 pub mod curves;
-/// Structs for points
-pub mod element;
 /// Pairing algorithms
 pub mod pairing;
+/// Structs for points
+pub mod point;
 /// Common behaviour for Elliptic curves.
 pub mod traits;
